Move exceptions to libertem.common
Import to `libertem.io.dataset.base.tiling_scheme.py` was added since
last release.
