+++ docs/source/changelog/misc/exceptions.rst
@@ -1,0 +1,5 @@
+[Misc] Moved exception classes to :mod:`libertem.common`
+========================================================
+* Moved exception classes to :mod:`libertem.common` for MIT
+  license compatibility in dataset code. They are re-exported to
+  the old import location for backwards compatibility (:pr:`1543`).
