+++ src/libertem/exceptions.py
@@ -1,21 +1,8 @@
-class UDFException(Exception):
-    """
-    Raised when the UDF interface is somehow misused
-    """
-    pass
+# Was moved to common for MIT license compatibility
+# in dataset code
 
+from libertem.common.exceptions import (
+    UDFException, ExecutorSpecException, UDFRunCancelled
+)
 
-class ExecutorSpecException(Exception):
-    """
-    Raised when there is an error specifying an
-    Executor or its resources / workers
-    """
-    pass
-
-
-class UDFRunCancelled(Exception):
-    """
-    Raised when the UDF run was cancelled, either when the job was cancelled
-    using :meth:`AsyncJobExecutor.cancel`, or when the underlying data source
-    was interrupted.
-    """
+__all__ = ["UDFException", "ExecutorSpecException", "UDFRunCancelled"]
